requirements.txt
torch>=1.13.1
transformers>=4.30.0
wandb>=0.15.3
requests>=2.28.1
huggingface-hub>=0.14.1
python-dotenv>=1.0.0
